Tests: Use better domain for HTTP API invalid hostname test.
The previous domain that was used to test for a host whose IPv4 address cannot be resolved, `exampleeeee.com`, got registered and has an A-record now, so it's not invalid anymore.

`.invalid` is intended for use in online construction of domain names that are sure to be invalid and which it is obvious at a glance are invalid.

Reference: [https://datatracker.ietf.org/doc/html/rfc2606#section-2 Reserved Top Level DNS Names: TLDs for Testing, & Documentation Examples].

Follow-up to [52084], [58384], [58388].

Props sippis, johnbillion, MattyRob, swissspidy.
Fixes #62303.
Built from https://develop.svn.wordpress.org/trunk@59293


git-svn-id: https://core.svn.wordpress.org/trunk@58685 1a063a9b-81f0-0310-95a4-ce76da25c4cd